+++ docs/stories/04-components/Dialog.stories.tsx
@@ -13,7 +13,7 @@ const meta: Meta<DialogArgs> = {
   title: 'Components/Dialog',
   decorators: [
     (Story) => (
-      <Flex style={{ width: '1280px', height: '720px' }} justifyContent="center">
+      <Flex width="100%" height="100%" justifyContent="center">
         <Story />
       </Flex>
     ),