+++ packages/design-system/src/components/Dialog/Dialog.tsx
@@ -102,15 +102,7 @@ interface HeaderProps extends TypographyProps<'h2'> {}
 const Header = React.forwardRef<HeaderElement, HeaderProps>(({ children, ...restProps }, forwardedRef) => {
   return (
     <AlertDialog.Title asChild>
-      <Title
-        tag="h2"
-        ref={forwardedRef}
-        padding={6}
-        justifyContent="center"
-        variant="h2"
-        fontWeight="bold"
-        {...restProps}
-      >
+      <Title<'h2'> tag="h2" variant="beta" ref={forwardedRef} padding={6} fontWeight="bold" {...restProps}>
         {children}
       </Title>
     </AlertDialog.Title>
@@ -189,7 +181,7 @@ const Description = React.forwardRef<DescriptionElement, DescriptionProps>((prop
 
 type FooterElement = HTMLDivElement;
 
-interface FooterProps {}
+interface FooterProps extends Omit<FlexProps<'footer'>, 'tag'> {}
 
 const Footer = React.forwardRef<FooterElement, FooterProps>((props, forwardedRef) => {
   return <Foot ref={forwardedRef} gap={2} padding={4} justifyContent="space-between" {...props} tag="footer" />;