+++ packages/design-system/src/components/Select/SelectParts.tsx
@@ -16,6 +16,8 @@ import { Typography, TypographyComponent, TypographyProps } from '../Typography'
  * SelectTrigger
  * -----------------------------------------------------------------------------------------------*/
 
+type TriggerSize = 'S' | 'M';
+
 interface TriggerProps extends BoxProps<'div'>, Pick<Field.InputProps, 'name' | 'id'> {
   /**
    * @default "Clear"
@@ -24,11 +26,19 @@ interface TriggerProps extends BoxProps<'div'>, Pick<Field.InputProps, 'name' | 
   disabled?: boolean;
   hasError?: boolean;
   onClear?: (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement>) => void;
+  /**
+   * @default "M"
+   */
+  size?: TriggerSize;
   startIcon?: React.ReactElement;
+  withTags?: boolean;
 }
 
 const SelectTrigger = React.forwardRef<HTMLDivElement, TriggerProps>(
-  ({ onClear, clearLabel = 'Clear', startIcon, disabled, hasError, children, id, ...restProps }, ref) => {
+  (
+    { onClear, clearLabel = 'Clear', startIcon, disabled, hasError, children, id, size = 'M', withTags, ...restProps },
+    ref,
+  ) => {
     const triggerRef = React.useRef<HTMLSpanElement>(null!);
 
     const handleClearClick = (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement>) => {
@@ -54,13 +64,11 @@ const SelectTrigger = React.forwardRef<HTMLDivElement, TriggerProps>(
           overflow="hidden"
           hasRadius
           background={disabled ? 'neutral150' : 'neutral0'}
-          paddingLeft={3}
-          paddingRight={3}
-          paddingTop={2}
-          paddingBottom={2}
           gap={4}
           cursor="default"
           aria-labelledby={labelNode ? `${id}-label` : undefined}
+          $size={size}
+          $withTags={withTags}
           {...restProps}
         >
           <Flex flex="1" tag="span" gap={3}>
@@ -114,8 +122,26 @@ const IconBox = styled<BoxComponent<'button'>>(Box)`
 
 const StyledTrigger = styled<FlexComponent>(Flex)<{
   $hasError?: boolean;
+  $size: TriggerSize;
+  $withTags?: boolean;
 }>`
   border: 1px solid ${({ theme, $hasError }) => ($hasError ? theme.colors.danger600 : theme.colors.neutral200)};
+  ${(props) => {
+    switch (props.$size) {
+      case 'S':
+        return css`
+          padding-block: ${props.theme.spaces[1]};
+          padding-inline-start: ${props.$withTags ? props.theme.spaces[1] : props.theme.spaces[4]};
+          padding-inline-end: ${props.theme.spaces[3]};
+        `;
+      default:
+        return css`
+          padding-block: ${props.$withTags ? '0.3rem' : props.theme.spaces[2]};
+          padding-inline-start: ${props.$withTags ? props.theme.spaces[1] : props.theme.spaces[4]};
+          padding-inline-end: ${props.theme.spaces[3]};
+        `;
+    }
+  }}
 
   &[aria-disabled='true'] {
     color: ${(props) => props.theme.colors.neutral600};
@@ -173,8 +199,7 @@ const SelectContent = styled(Select.Content)`
   border: 1px solid ${({ theme }) => theme.colors.neutral150};
   border-radius: ${({ theme }) => theme.borderRadius};
   min-width: var(--radix-select-trigger-width);
-  /* This is from the design-system figma file. */
-  max-height: 15rem;
+  max-height: 15.6rem;
   z-index: ${({ theme }) => theme.zIndices.popover};
 
   @media (prefers-reduced-motion: no-preference) {