+++ packages/primitives/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @strapi/ui-primitives
 
+## 2.0.0-beta.6
+
+## 2.0.0-beta.5
+
 ## 2.0.0-beta.4
 
 ## 2.0.0-beta.3